Testing attention maps in gloria

+++ contrastive_training.py
@@ -230,7 +230,7 @@ def contrastive_pretraining(config):
             for i in range(0, 8):
                 print("iteration thingy: " + str(i))
                 #print(attn_maps[i])
-                im = np.uint8(attn_maps[i].squeeze().cpu().detach().numpy())
+                im = attn_maps[i].squeeze().cpu().detach().numpy()
                 print(im)
                 print(type(im))
                 im = Image.fromarray(im)
